move main method to new class
